taking screenshot and share picture with exercise data works, but picture is all black. looking into it....

+++ app/src/main/java/com/uniks/myfit/database/LocationDataDao.java
@@ -15,7 +15,7 @@ public interface LocationDataDao {
     List<LocationData> getAll();
 
     @Query("SELECT * FROM locationdata WHERE exercise_id = :exerciseId")
-    List<LocationData> getAllFromExercise(int exerciseId);
+    List<LocationData> getAllFromExercise(long exerciseId);
 
     @Update
     void updateLocationDatas(LocationData... locationData);
